Cleanup and pull "not_after" out of the cert.
This is a little cleanup mostly to read "not_after" out of the certificate, in order to manage certificate expiration generally.  
The general use case is just certificates expiring soon, as below:

    $soon = (Get-date).AddDays(21)
    Get-CertTransparencyInfo -SearchInfo "%.example.com" -GetCertificate | where-object {$_.not_after -lt $soon} | format-table -Property name_value, not_after, issuer_name

Another use case would be to identify identify certificates impacted by the Symantec root deprecation, as documented below.
https://arkadiyt.com/2018/02/04/quantifying-untrusted-symantec-certificates/

    $m70 = Get-Date -Date "2018-09-01 00:00:00Z"
    Get-CertTransparencyInfo -SearchInfo "%.example.com" -GetCertificate | where-object {$_.not_after -gt $m70 -and $_.issuer_name -like "*Symantec*"} | sort-object -property not_after | format-table -Property name_value, not_after, issuer_name

I realize people need to look for other Symantec brand issuers like  Thawte, Geotrust, and RapidSSL.  I'm working on a more detailed version of the above, which would then check if the hosts are still using the deprecated certificate.